Cast

+++ FrameLib_PD_Objects/Common/FrameLib_PDClass.h
@@ -1362,7 +1362,7 @@ private:
                 else
                 {
                     i = parseNumericalList(values, argv, argc, i);
-                    serialisedParameters.write(sym->s_name + 1, values.data(), values.size());
+                    serialisedParameters.write(sym->s_name + 1, values.data(), static_cast<unsigned long>(values.size()));
                 }
             }
         }
